+++ src/main/java/com/exemplo/projeto/perfil_web/repository/ServicoRepository.java
@@ -7,6 +7,6 @@ package com.exemplo.projeto.perfil_web.repository;
 import com.exemplo.projeto.perfil_web.model.Servico;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServicoRepository extends JpaRepository<Servico, Long>{
+public interface ServicoRepository extends JpaRepository<Servico, Integer>{
     
 }
